+++ recherche_note.h
@@ -7,6 +7,6 @@ unsigned int shift(unsigned int indice, unsigned int longueur, double* abs_tabTF
 
 double argmax(double fe, unsigned int longueur, double* abs_tabTF);
 
-void recherche_note(double fe, unsigned int longueur, double* abs_tabTF, unsigned int* pcolor, unsigned int* ppos, char* note, int* pf, unsigned int* poctave);
+void recherche_note(double fe, unsigned int longueur, double* abs_tabTF, unsigned int* pcolor, unsigned int* ppos, char* note, double* pf, unsigned int* poctave);
 
 #endif
